{"schema":"decodex/commit/1","summary":"Split operator project display status","authority":"XY-1115"}
